refactor: rename ingress-invite-store.ts to invite-store.ts
